ScreenMonitor is for MozStumbler use, not Fennec. Move to client side.
